Clarify need for `WallSegment` script
The broken wall and solid wall segments in sieges seem to need a `WallSegment` script for the map not to crash (default values suffice).

When following the guide, I did not add these scripts to my custom prefab as it's not mentioned anywhere. I hope with this clarification, someone else's headache can be saved.